+++ rss-puzzle/src/components/main/eventHandlers.ts
@@ -1,11 +1,10 @@
-export function setupEventHandlers(){
+export function setupEventHandlers() {
+  const resultBlock = document.getElementById('result-block');
 
-const resultBlock = document.getElementById('result-block');
+  resultBlock.id = 'result-block';
+  resultBlock.classList.add('container');
 
-resultBlock.id = 'result-block';
-resultBlock.classList.add('container');
-
-const nextButton = document.getElementById(
+  const nextButton = document.getElementById(
     'next-sentence-button',
   ) as HTMLButtonElement;
   nextButton.disabled = true;
@@ -25,21 +24,18 @@ const nextButton = document.getElementById(
   nextButton.style.visibility = 'hidden';
 
   document
-  .getElementById('toggle-translation-button')
-  .addEventListener('click', () => {
-    const translationSpan = document.querySelector(
-      '.translation-hint',
-    ) as HTMLElement;
-    const button = document.getElementById(
-      'toggle-translation-button',
-    ) as HTMLButtonElement;
-    if (translationSpan) {
-      const isVisible = translationSpan.style.display !== 'none';
-      translationSpan.style.display = isVisible ? 'none' : 'block';
-      button.textContent = isVisible
-        ? 'Показать перевод'
-        : 'Скрыть перевод';
-    }
-  });
-
+    .getElementById('toggle-translation-button')
+    .addEventListener('click', () => {
+      const translationSpan = document.querySelector(
+        '.translation-hint',
+      ) as HTMLElement;
+      const button = document.getElementById(
+        'toggle-translation-button',
+      ) as HTMLButtonElement;
+      if (translationSpan) {
+        const isVisible = translationSpan.style.display !== 'none';
+        translationSpan.style.display = isVisible ? 'none' : 'block';
+        button.textContent = isVisible ? 'Показать перевод' : 'Скрыть перевод';
+      }
+    });
 }
