+++ rss-puzzle/src/components/main/fetchData.ts
@@ -1,7 +1,9 @@
 import { IWordData } from './interfaces';
 
-export async function fetchWordData(dataUrl: string): Promise<IWordData | undefined> {
- try {
+export async function fetchWordData(
+  dataUrl: string,
+): Promise<IWordData | undefined> {
+  try {
     const response = await fetch(dataUrl);
     if (!response.ok) {
       throw new Error(`HTTP error! status: ${response.status}`);
@@ -9,7 +11,7 @@ export async function fetchWordData(dataUrl: string): Promise<IWordData | undefi
     const wordData = await response.json();
     console.log('Fetched data:', wordData);
     return wordData;
- } catch (error) {
+  } catch (error) {
     console.error('Error fetching word data:', error);
- }
+  }
 }
